KNIME-1705 FunctionalGroupFilter node: migration to KNIME NIO (#135)
Authored-by: Roman Balabanov <roman-1.balabanov_ext@novartis.com>
