+++ org.rdkit.knime.nodes/src-deprecated/org/rdkit/knime/nodes/functionalgroupfilter/FunctionalGroupFilterNodeFactory.xml
