--- org.rdkit.knime.nodes/src/org/rdkit/knime/nodes/functionalgroupfilter/FunctionalGroupFilterNodeFactory.java
@@ -1,117 +1,0 @@
-/*
- * ------------------------------------------------------------------
- * This source code, its documentation and all appendant files
- * are protected by copyright law. All rights reserved.
- *
- * Copyright (C) 2012
- * Novartis Institutes for BioMedical Research
- *
- *
- *  This program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License, Version 3, as
- *  published by the Free Software Foundation.
- *
- *  This program is distributed in the hope that it will be useful, but
- *  WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program; if not, see <http://www.gnu.org/licenses>.
- *
- *  Additional permission under GNU GPL version 3 section 7:
- *
- *  KNIME interoperates with ECLIPSE solely via ECLIPSE's plug-in APIs.
- *  Hence, KNIME and ECLIPSE are both independent programs and are not
- *  derived from each other. Should, however, the interpretation of the
- *  GNU GPL Version 3 ("License") under any applicable laws result in
- *  KNIME and ECLIPSE being a combined program, KNIME GMBH herewith grants
- *  you the additional permission to use and propagate KNIME together with
- *  ECLIPSE with only the license terms in place for ECLIPSE applying to
- *  ECLIPSE and the GNU GPL Version 3 applying for KNIME, provided the
- *  license terms of ECLIPSE themselves allow for the respective use and
- *  propagation of ECLIPSE together with KNIME.
- *
- *  Additional permission relating to nodes for KNIME that extend the Node
- *  Extension (and in particular that are based on subclasses of NodeModel,
- *  NodeDialog, and NodeView) and that only interoperate with KNIME through
- *  standard APIs ("Nodes"):
- *  Nodes are deemed to be separate and independent programs and to not be
- *  covered works.  Notwithstanding anything to the contrary in the
- *  License, the License does not apply to Nodes, you are not required to
- *  license Nodes under the License, and you are granted a license to
- *  prepare and propagate Nodes, in each case even if such Nodes are
- *  propagated with or for interoperation with KNIME.  The owner of a Node
- *  may freely choose the license terms applicable to such Node, including
- *  when such Node is propagated with or for interoperation with KNIME.
- * ---------------------------------------------------------------------
- */
-package org.rdkit.knime.nodes.functionalgroupfilter;
-
-import org.knime.core.node.NodeDialogPane;
-import org.knime.core.node.NodeFactory;
-import org.knime.core.node.NodeView;
-
-/**
- * <code>NodeFactory</code> for the RDKit based "RDKitFunctionalGroupFilter" Node.
- * 
- * @author Dillip K Mohanty
- * @author Manuel Schwarze
- */
-public class FunctionalGroupFilterNodeFactory extends NodeFactory<FunctionalGroupFilterNodeModel> {
-
-	/**
-	 * Creates a model for the RDKitFunctionalGroupFilter functionality
-	 * of the RDKit library. The model is derived from the
-	 * abstract class AbstractRDKitNodeModel, which provides
-	 * common base functionality for RDKit nodes.
-	 * {@inheritDoc}
-	 *
-	 * @see org.rdkit.knime.nodes.AbstractRDKitNodeModel
-	 */
-	@Override
-	public FunctionalGroupFilterNodeModel createNodeModel() {
-		return new FunctionalGroupFilterNodeModel();
-	}
-
-	/**
-	 * This node does not have any views.
-	 * 
-	 * @return Always null.
-	 */
-	@Override
-	public NodeView<FunctionalGroupFilterNodeModel> createNodeView(
-			final int viewIndex,
-			final FunctionalGroupFilterNodeModel nodeModel) {
-		return null;
-	}
-
-	/**
-	 * This node does not have any views.
-	 * 
-	 * @return Always 0.
-	 */
-	@Override
-	public int getNrNodeViews() {
-		return 0;
-	}
-
-	/**
-	 * This node possesses a configuration dialog.
-	 * 
-	 * @return Always true.
-	 */
-	@Override
-	public boolean hasDialog() {
-		return true;
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public NodeDialogPane createNodeDialogPane() {
-		return new FunctionalGroupFilterNodeDialog();
-	}
-}
-
